Fix bug with dormant on multiple pins
There was a bug when going dormant on multiple pins would only work if
all the pins where on the same wake register. Otherwise only the
register for the last handled pin would be set, causing some of the pins
to not be registered.

Change the code so that each wake register is not set properly.